Get Delete working again (even if temp)
