+++ handlers/org.js
@@ -12,31 +12,5 @@ exp.getOrg = function (request, reply) {
   });
 };
 
-exp.updateOrg = function (request, reply) {
-  var opts = {};
-  var loggedInUser = request.loggedInUser.name;
-  var orgName = request.params.org;
-  var user = {
-    user: request.payload.username,
-    role: request.payload.role
-  };
-  if (request.payload.updateType === "addUser") {
-  } else if (request.payload.updateType === "deleteUser") {
-    Org(loggedInUser)
-      .removeUser(orgName, user.user, function (err, removedUser) {
-        if (err) {
-          request.logger.error(err);
-          return reply.view('errors/internal', err);
-        }
-        Org(loggedInUser)
-          .get(orgName, function (err, org) {
-            if (err) { request.logger.error(err); }
-            opts.org = org;
-            return reply.view('org/info', opts);
-          });
-      });
-  }
-};
-
 
 module.exports = exp;
